chore: Update tsconfig.json and package.json
The tsconfig.json file was updated to include the "src/components/Spotlightbutton.tsx" file in the include section.

The package.json file was updated to add new dependencies:
- "@radix-ui/react-direction": "^1.0.1"
- "@radix-ui/react-navigation-menu": "^1.1.4"
- "@radix-ui/themes": "^3.0.5"
- "@studio-freight/react-lenis": "^0.0.47"
- "framer-motion": "^11.2.9"
- "lenis": "^1.0.45"
- "react": "^18.3.1"
- "react-dom": "^18.3.1"

The devDependencies section of package.json was also updated to include:
- "eslint": "^8"
- "eslint-config-next": "14.2.3"

The commit message suggests that these changes were made to update the project's configuration and dependencies.